requirements.txt
langchain==0.0.352
zhipuai==1.0.7
python-dotenv==1.0.0
faiss-cpu==1.7.4
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
urllib3<2.0.0
flask==3.0.0
flask-socketio==5.3.6
SpeechRecognition==3.10.0
python-docx==1.0.0
markdown==3.5.1
chromadb==0.4.22
PyAudio==0.2.13
